requirements.txt
# AI-First CRM Backend
# FastAPI with intelligent ticket management

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.1.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# API & HTTP
httpx==0.28.0
aiohttp==3.11.7
python-multipart==0.0.18

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI & ML
openai==1.56.0
anthropic==0.39.0
tiktoken==0.8.0

# Background Jobs
celery==5.4.0
redis==5.2.0

# Data Processing
pandas==2.2.3
numpy==2.2.0
python-dateutil==2.8.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
